requirements.txt
requests>=2.21.0
Click>=7.0
click_completion
termcolor>=1.1.0
Flask-Bootstrap4
